Change multiply_numbers to accept double parameters

+++ src/homework/01_data_types/data_types.cpp
@@ -1,12 +1,8 @@
 // data_types.cpp
 #include "data_types.h"
 
-int multiply_numbers(int num1, int num2)
+// Function definition
+double multiply_numbers(double num1, double num2)
 {
     return num1 * num2;
 }
-
-
-
-
-
